Refactor Mobile Setting by removing MobileUpdateForm and instead reuse MobileForm for the common operation
- Implementation for firebase still in progress
