+++ src/pages/Settings/Authentication/index.tsx
@@ -4,7 +4,6 @@ import { Card } from "@/elements";
 import React from "react";
 import { LoginEmailForm } from "@/components/LoginEmailForm";
 import { LoginPasswordForm } from "@/components/LoginPasswordForm";
-import { MobileAuthSection } from "./MobileAuthSection";
 import { MobileNumberSetting } from "@/components/MobileNumberSetting";
 
 export const SettingAuthenticationPage: React.FC = () => {
@@ -29,7 +28,6 @@ export const SettingAuthenticationPage: React.FC = () => {
         <LoginPasswordForm />
       </SectionContainer>
       <MobileNumberSetting />
-      {/* <MobileAuthSection /> */}
     </VStack>
   );
 };
